+++ lending-backend-dotnet/src/LendingSystem.Auth.Application/Abstractions/IUserRepository.cs
@@ -9,12 +9,12 @@ public interface IUserCommandRepository
     Task<UserEntity?> FindByProviderAsync(AuthProvider authProvider, string providerUserId, CancellationToken cancellationToken);
     Task<UserProfile> CreateAsync(string name, string email, string passwordHash, CancellationToken cancellationToken);
     Task<UserEntity> CreateExternalAsync(string name, string email, AuthProvider authProvider, string providerUserId, CancellationToken cancellationToken);
-    Task<UserEntity?> LinkProviderAsync(int userId, AuthProvider authProvider, string providerUserId, CancellationToken cancellationToken);
-    Task<bool> DeleteAsync(int userId, CancellationToken cancellationToken);
+    Task<UserEntity?> LinkProviderAsync(long userId, AuthProvider authProvider, string providerUserId, CancellationToken cancellationToken);
+    Task<bool> DeleteAsync(long userId, CancellationToken cancellationToken);
 }
 
 public interface IUserQueryRepository
 {
-    Task<UserProfile?> GetByIdAsync(int userId, CancellationToken cancellationToken);
+    Task<UserProfile?> GetByIdAsync(long userId, CancellationToken cancellationToken);
     Task<UserProfile?> SearchByNameAsync(string username, CancellationToken cancellationToken);
 }
