+++ lending-backend-dotnet/src/LendingSystem.Auth.Application/Auth/DeleteUser/DeleteUserCommand.cs
@@ -3,4 +3,4 @@ using MediatR;
 
 namespace LendingSystem.Auth.Application.Auth;
 
-public sealed record DeleteUserCommand(int UserId) : ICommand<DeleteUserResult>;
+public sealed record DeleteUserCommand(long UserId) : ICommand<DeleteUserResult>;
