+++ lending-backend-dotnet/src/LendingSystem.Auth.Application/Auth/GetUserById/GetUserByIdResult.cs
@@ -3,6 +3,6 @@ using System.Text.Json.Serialization;
 namespace LendingSystem.Auth.Application.Auth;
 
 public sealed record GetUserByIdResult(
-    [property: JsonIgnore] int UserId,
+    [property: JsonIgnore] long UserId,
     [property: JsonPropertyName("name")] string Name,
     [property: JsonPropertyName("email")] string Email);
