+++ lending-backend-dotnet/src/LendingSystem.SharedKernel.Application/Abstractions/IExecutionContextAccessor.cs
@@ -3,9 +3,9 @@ namespace LendingSystem.SharedKernel.Application.Abstractions;
 public interface IExecutionContextAccessor
 {
     Guid UserId { get; }
-    int CurrentUserId { get; }
+    long CurrentUserId { get; }
     string Email { get; }
     string PasswordHash { get; }
     bool IsAdmin { get; }
-    bool CanAccessUser(int userId);
+    bool CanAccessUser(long userId);
 }
